Fix some bugs
No longer read whole bulks before splitting partitions. But it requires
txids of the log are in a sequence.
Parameter Firsttxid can be set to start a stream in any legit location.
Fix some bugs about Hash in partitions.

Todo: create a writer to get syslog and put it into a stream.